Move input/output functions to io object
